code/requirements.txt
# Seq2Seq模型实验依赖
# Python 3.11
# 适用于 Ubuntu 22.04 + NVIDIA GPU

# 核心深度学习框架
torch==2.7.1
torchvision>=0.18.0
torchaudio>=2.3.0

# Hugging Face相关
transformers>=4.36.0
datasets>=2.14.6
tokenizers>=0.15.0
huggingface-hub>=0.20.0

# NLP工具
spacy>=3.7.2
nltk>=3.8.1
sacremoses>=0.1.1

# 评估指标
rouge-score>=0.1.2
bert-score>=0.3.13
sacrebleu>=2.3.0

# 数据处理与科学计算
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0

# 配置与日志
PyYAML>=6.0
tqdm>=4.66.0

# 其他工具
pillow>=10.0.0
sentencepiece>=0.1.99
protobuf>=4.24.0

# 下载spaCy模型（运行时自动下载）
# python -m spacy download en_core_web_sm
# python -m spacy download de_core_news_sm
